remove double quotes 🤦 and remove processing of `omitFields`
We cannot process `omitFields` here, unless by utterly complicating
these utilities by relying on TypeORM internals.

We could use set difference `fields` - `omitFields` only if `fields` is
populated in the request, but this would be quite unlikely: why would an
API consumer normally provide a field in `fields` and then ask for it to
be omitted in `omitFields`?

So basically in the case where a user only wants to omit one or more
fields via `omitFields`, we would have to use the TypeORM API to get a
list of all the fields that would be selected, and then subtract from
this set anything which is in `omitFields`. Way too brittle.

What we'll do instead is to remove fields listed in `omitFields` *after*
fetching and before returning the object to the caller, in the default
BaseService request lifecycle.

This isn't ideal either, but should be a decent tradeoff.

In case of large omittable fields (for example, GeoJSON data from
geometry columns) we'd end up querying and receiving over the wire data
that we'd then discard.

Moreover, despite this being potentially brittle in practice because
we'd have to work against the TypeORM grain (see
https://github.com/typeorm/typeorm/issues/535), it *would* indeed be
cleaner to deal with `omitFields` in the same place where we deal with
`fields`.